use `SimulationControls` as single source of truth
To prepare for encoding the control values into the URL, it makes our lives a whole lot easier to have a single source of truth of those values. I'd say it also cleans op initialization a bit. The only problem i have with this is that some places in the code where we get the control values it becomes *very* verbose.